requirements.txt
# Core dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

# Data processing
pandas==2.1.4
numpy==1.24.3

# Scheduling
schedule==1.2.0
APScheduler==3.10.4

# Logging and monitoring
loguru==0.7.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: For data visualization (if needed for reports)
matplotlib==3.8.2
seaborn==0.13.0
